New Date Input
 - Code and Date input handling moved from xibo-cms to the new forms
