+++ modules/embedded.xml
@@ -49,14 +49,14 @@
             <helpText>Should this Widget be loaded entirely off-screen so that it is ready when shown? Dynamic content will start running off screen.</helpText>
             <default>1</default>
         </property>
-        <property id="embedHtml" type="code" allowLibraryRefs="true" codeType="html">
+        <property id="embedHtml" type="code" allowLibraryRefs="true" variant="html">
             <title>HTML to embed</title>
             <helpText>Enter text or HTML in the box below.</helpText>
         </property>
-        <property id="embedStyle" type="code" allowLibraryRefs="true" codeType="css">
+        <property id="embedStyle" type="code" allowLibraryRefs="true" variant="css">
             <title>Custom Style Sheets</title>
         </property>
-        <property id="embedScript" type="code" allowLibraryRefs="true" codeType="javascript">
+        <property id="embedScript" type="code" allowLibraryRefs="true" variant="javascript">
             <title>JavaScript to embed</title>
             <helpText>Please include script tags.</helpText>
         </property>
