+++ ui/src/templates/forms/inputs/code.hbs
@@ -3,7 +3,7 @@
     {{#if helpText}}
         <i class="fa fa-question-circle form-help-popup" data-toggle="tooltip" data-placement="top" title="{{helpText}}"></i>
     {{/if}}
-    <textarea class="form-control code-input d-none" id="{{id}}" name="{{id}}" rows="{{rows}}" data-code-type="{{codeType}}">{{value}}</textarea>
+    <textarea class="form-control code-input d-none" id="{{id}}" name="{{id}}" rows="{{rows}}" data-code-type="{{variant}}">{{value}}</textarea>
     <div class="code-input-editor-container">
         <div class="code-input-editor"></div>
     </div>
