+++ ui/src/templates/forms/inputs/date.hbs
@@ -1,0 +1,12 @@
+<div class="form-group xibo-form-input date-input-group" {{#if visibility}}data-visibility="{{visibility}}" {{/if}}>
+    <label for="{{id}}" class="control-label"><strong>{{title}}</strong></label>
+    {{#if helpText}}
+        <i class="fa fa-question-circle form-help-popup" data-toggle="tooltip" data-placement="top"
+            title="{{helpText}}"></i>
+    {{/if}}
+    <div class="input-group">
+        <input type="text" class="form-control dateControl {{variant}}" id="{{id}}" name="{{id}}" value="{{value}}" {{#if format }}data-custom-format="{{ format }}"{{/if}} />
+        <span class="input-group-append input-group-addon input-group-text date-clear-button d-none" role="button"><i
+                class="fa fa-times"></i></span>
+    </div>
+</div>
