+++ views/include-file-upload.twig
@@ -173,7 +173,7 @@
                     <div class="col-sm-10">
                         <div class="input-group">
                             <span class="input-group-prepend input-group-text date-open-button" role="button"><i class="fa fa-calendar"></i></span>
-                            <input class="form-control dateTimePicker dateControl" type="text" name="widgetFromDt" id="widgetFromDt" />
+                            <input class="form-control dateControl dateTime" type="text" name="widgetFromDt" id="widgetFromDt" />
                             <span class="input-group-append input-group-addon input-group-text date-clear-button d-none" role="button"><i class="fa fa-times"></i></span>
                         </div>
                        <small class="form-text text-muted">{{ trans.widgetFromDtMessage }}</small>
@@ -186,7 +186,7 @@
                     <div class="col-sm-10">
                         <div class="input-group">
                             <span class="input-group-prepend input-group-text date-open-button" role="button"><i class="fa fa-calendar"></i></span>
-                            <input class="form-control dateTimePicker dateControl" type="text" name="widgetToDt" id="widgetToDt" />
+                            <input class="form-control dateControl dateTime" type="text" name="widgetToDt" id="widgetToDt" />
                             <span class="input-group-append input-group-addon input-group-text date-clear-button d-none" role="button"><i class="fa fa-times"></i></span>
                         </div>
                        <small class="form-text text-muted">{{ trans.widgetToDtMessage }}</small>
